商品削除機能の修正2

+++ app/controllers/items_controller.rb
@@ -40,9 +40,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    return unless item.user_id == current_user.id
-
-    item.destroy
+    item.destroy if item.user_id == current_user.id
     redirect_to root_path
   end
